Refactor in namespace and files.
