+++ Elements/src/Geometry/Tessellation/Tessellation.cs
@@ -1,189 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-using Elements.Geometry.Solids;
 using LibTessDotNet.Double;
 [assembly: InternalsVisibleTo("Hypar.Elements.Tests")]
 
-namespace Elements.Geometry
+namespace Elements.Geometry.Tessellation
 {
-    /// <summary>
-    /// An object which provides tessellation targets.
-    /// </summary>
-    internal interface ITessellationTargetProvider
-    {
-        /// <summary>
-        /// Get the tessellation targets.
-        /// </summary>
-        IEnumerable<ITtessAdator> GetTessellationTargets();
-    }
-
-    /// <summary>
-    /// An object which provides tessellation targets for a solid.
-    /// </summary>
-    internal class SolidTesselationTargetProvider : ITessellationTargetProvider
-    {
-        private readonly Solid solid;
-        private readonly Transform transform;
-
-        /// <summary>
-        /// Construct a SolidTesselationTargetProvider.
-        /// </summary>
-        /// <param name="solid"></param>
-        /// <param name="transform"></param>
-        public SolidTesselationTargetProvider(Solid solid, Transform transform = null)
-        {
-            this.solid = solid;
-            this.transform = transform;
-        }
-
-        /// <summary>
-        /// Get the tessellation targets.
-        /// </summary>
-        public IEnumerable<ITtessAdator> GetTessellationTargets()
-        {
-            foreach (var f in solid.Faces.Values)
-            {
-                yield return new SolidFaceTessAdaptor(f, transform);
-            }
-        }
-    }
-
-    /// <summary>
-    /// An object which provides tessellation targets for a csg solid.
-    /// </summary>
-    internal class CsgTessellationTargetProvider : ITessellationTargetProvider
-    {
-        private readonly Csg.Solid csg;
-
-        /// <summary>
-        /// Construct a CsgTessellationTargetProvider.
-        /// </summary>
-        /// <param name="csg"></param>
-        public CsgTessellationTargetProvider(Csg.Solid csg)
-        {
-            this.csg = csg;
-        }
-
-        /// <summary>
-        /// Get the tessellation targets.
-        /// </summary>
-        public IEnumerable<ITtessAdator> GetTessellationTargets()
-        {
-            foreach (var p in csg.Polygons)
-            {
-                yield return new CsgPolygonTessAdaptor(p);
-            }
-        }
-    }
-
-    /// <summary>
-    /// An object which creates a tessellation for a tessellation target.
-    /// </summary>
-    internal interface ITtessAdator
-    {
-        /// <summary>
-        /// Does this target require tessellation?
-        /// </summary>
-        bool RequiresTessellation();
-
-        /// <summary>
-        /// Get the tessellation.
-        /// </summary>
-        Tess GetTess();
-    }
-
-    /// <summary>
-    /// An object which provides a tessellation for a csg polygon.
-    /// </summary>
-    internal class CsgPolygonTessAdaptor : ITtessAdator
-    {
-        private readonly Csg.Polygon polygon;
-
-        /// <summary>
-        /// Construct a CsgPolygonTessAdaptor.
-        /// </summary>
-        /// <param name="polygon"></param>
-        public CsgPolygonTessAdaptor(Csg.Polygon polygon)
-        {
-            this.polygon = polygon;
-        }
-
-        /// <summary>
-        /// Get the tessellation.
-        /// </summary>
-        public Tess GetTess()
-        {
-            var tess = new Tess
-            {
-                NoEmptyPolygons = true
-            };
-            tess.AddContour(polygon.Vertices.ToContourVertices());
-
-            tess.Tessellate(WindingRule.Positive, ElementType.Polygons, 3);
-            return tess;
-        }
-
-        /// <summary>
-        /// Does this target require tessellation?
-        /// </summary>
-        public bool RequiresTessellation()
-        {
-            return polygon.Vertices.Count > 3;
-        }
-    }
-
-    /// <summary>
-    /// An object which provides a tessellation for a solid face.
-    /// </summary>
-    internal class SolidFaceTessAdaptor : ITtessAdator
-    {
-        private readonly Face face;
-        private readonly Transform transform;
-
-        /// <summary>
-        /// Construct a SolidFaceTessAdaptor.
-        /// </summary>
-        /// <param name="face"></param>
-        /// <param name="transform"></param>
-        public SolidFaceTessAdaptor(Face face, Transform transform = null)
-        {
-            this.face = face;
-            this.transform = transform;
-        }
-
-        /// <summary>
-        /// Get the tessellation.
-        /// </summary>
-        public Tess GetTess()
-        {
-            var tess = new Tess
-            {
-                NoEmptyPolygons = true
-            };
-            tess.AddContour(face.Outer.ToContourVertexArray(transform));
-
-            if (face.Inner != null)
-            {
-                foreach (var loop in face.Inner)
-                {
-                    tess.AddContour(loop.ToContourVertexArray(transform));
-                }
-            }
-
-            tess.Tessellate(WindingRule.Positive, ElementType.Polygons, 3);
-            return tess;
-        }
-
-        /// <summary>
-        /// Does this target require tessellation?
-        /// </summary>
-        public bool RequiresTessellation()
-        {
-            return true;
-        }
-    }
-
     /// <summary>
     /// Methods for the tessellation of various objects.
     /// </summary>
